Presenter 2.14.x

+++ app/presenter/datasource/controller/datasource-view.js
@@ -1,9 +1,10 @@
 define([
     'connecta.presenter',
     'presenter/datasource/service/datasource-service',
-    'presenter/datasource/filter/bullet'
+    'presenter/datasource/filter/bullet',
+    'portal/layout/service/confirm'
 ], function (presenter) {
-    return presenter.lazy.controller('DatasourceViewController', function ($scope, DatasourceService, $routeParams, $location, notify) {
+    return presenter.lazy.controller('DatasourceViewController', function ($scope, $confirm, DatasourceService, $routeParams, $location, notify) {
 
         DatasourceService.getById($routeParams.id).then(function (response) {
             $scope.datasource = response.data;
@@ -11,13 +12,15 @@ define([
             $scope.types = DatasourceService.getTypes();
 
             $scope.excluir = function (id) {
-                DatasourceService.remove(id).then(function () {
-                    $location.path('presenter/datasource');
+                $confirm('DATASOURCE.BULK_DELETE_CONFIRM', 'DATASOURCE.BULK_CONFIRM_DELETE').then(function(){
+                    DatasourceService.remove(id).then(function () {
+                        $location.path('presenter/datasource');
+                    });
                 });
             };
 
             $scope.testConnection = function () {
-                DatasourceService.testConnection($scope.datasource).then(function () {
+                    DatasourceService.testConnection($scope.datasource).then(function () {
                     notify.success('Conexão feita com sucesso.');
                 });
             };
